CyberVenom AI/requirements.txt
requests==2.31.0
python-nmap==0.7.1
PyQt5==5.15.9
cryptography>=41.0.5
pycryptodome==3.19.0
python-dotenv==1.0.0
pyOpenSSL>=24.0.0
